# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sharefy",
        "description": "",
        "contact": {
            "email": "equip@sharefy.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/comment/user/{username}/publication/{id_publication}": {
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "Crea un comentari a una publicació",
                "description": "Dado un username, un id publicació, un texto del comentari y la data actual del comentari, es crea un comentari",
                "operationId": "App\\Http\\Controllers\\Api\\CommentDataController::create",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Text del comentari a fer.",
                        "required": true
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el mensaje: 'Comentari crear correctament'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    },
                    "404": {
                        "description": "Devuelve un json con el error: 'Usuari no trobat'."
                    }
                }
            }
        },
...
        "/api/comments/publication/{id}": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "Devuelve los comentarios de una publicación.",
                "description": "Dado un id publicación, retorna un json con los comentarios y datos del usuario la publicación",
                "operationId": "App\\Http\\Controllers\\Api\\CommentDataController::get_comments",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json comments: [ {'id_comment', 'text', 'created_at', 'id_user', 'username',  'first_name', 'photo_path'} ]."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    }
                }
            }
        },
...
        "/api/like/user/{username}/publication/{id_publication}": {
            "get": {
                "tags": [
                    "like"
                ],
                "summary": "Comproba si un usuario ha dado like a una publicación",
                "description": "Dado un username y un id publicación retorna: 'true' o 'false', si el usuario ha dado like a esta publicación o no, respectivamente.",
                "operationId": "App\\Http\\Controllers\\Api\\LikeDataController::is_like",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el value: 'true' o 'false'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    },
                    "404": {
                        "description": "Devuelve un json con el error: 'usuari no trobat a la base de dades'.'"
                    }
                }
            },
            "post": {
                "tags": [
                    "like"
                ],
                "summary": "dado un username se crea la relacion like con la publicacion id_publicacion",
                "description": "dado un username se crea la relacion like con la publicacion id_publicacion",
                "operationId": "App\\Http\\Controllers\\Api\\LikeDataController::set_like",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el mensaje: 'relació creada correctament'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    }
                }
            },
            "delete": {
                "tags": [
                    "like"
                ],
                "summary": "dado un username se elimina la relacion like con la publicacion id_publicacion",
                "description": "dado un username se elimina la relacion like con la publicacion id_publicacion",
                "operationId": "App\\Http\\Controllers\\Api\\LikeDataController::unset_like",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el mensaje: 'relació eliminada correctament'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    }
                }
            }
        },
...
        "/api/publication/{id}": {
            "get": {
                "tags": [
                    "publication"
                ],
                "summary": "Dado un id de publicación existente, devuelve su información.",
                "description": "Dado un id de publicación existente, devuelve su información.",
                "operationId": "App\\Http\\Controllers\\Api\\PublicationDataController::show",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con la información de la publicación."
                    }
                }
            },
            "put": {
                "tags": [
                    "publication"
                ],
                "summary": "Dado un id de publicación existente, edita dicha publicación.",
                "description": "Dado un id de publicación existente, y los campos a modificados guarda la dicha informacioón.",
                "operationId": "App\\Http\\Controllers\\Api\\PublicationDataController::update",
                "parameters": [
                    {
                        "name": "game",
                        "in": "query",
                        "description": "Id del juego"
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Texto descriptivo de la publicación"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el mensaje: 'Publicació editada correctament'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'No existe la publicación a editar'."
                    }
                }
            }
        },
        "/api/user/{id_user}/publications": {
            "get": {
                "tags": [
                    "publication"
                ],
                "summary": "Dado un id de usuario existente, devuelve todas sus publicaciones",
                "description": "Dado un id de usuario existente, devuelve todas sus publicaciones",
                "operationId": "App\\Http\\Controllers\\Api\\PublicationDataController::list_user",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con la información de la publicación."
                    }
                }
            }
        },
        "/api/user/{id}/wall": {
            "get": {
                "tags": [
                    "publication"
                ],
                "summary": "Dado un id de usuario existente, devuelve las publicaciones más recientes de todos sus seguidos y las suyas (Muro).",
                "description": "Dado un id de usuario, devuelve toda la información de todas las publicaciones recientes que han hecho los usuarios al cual sigue y las suyas propias",
                "operationId": "App\\Http\\Controllers\\Api\\PublicationDataController::wall",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con la información de la publicación, users, likes, comentarios, game"
                    }
                }
            }
        },
        "/api/publication": {
            "post": {
                "tags": [
                    "publication"
                ],
                "summary": "Es crea la publicació amb la informació enviada.",
                "description": "Es crea la publicació amb la informació enviada.",
                "operationId": "App\\Http\\Controllers\\Api\\PublicationDataController::create",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "query",
                        "description": "string amb el id de l'usuari",
                        "required": true
                    },
                    {
                        "name": "game",
                        "in": "query",
                        "description": "id del game",
                        "required": true
                    },
                    {
                        "name": "video",
                        "in": "query",
                        "description": "valor del video en el form",
                        "required": true
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "string amb el text de la publicació",
                        "required": true
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna un json amb el identificador de la publicació"
                    },
                    "401": {
                        "description": "El vídeo no té un format vàlid, format suportat .mp4"
                    }
                }
            }
        },
        "/api/publication/{publication_id}": {
            "delete": {
                "tags": [
                    "publication"
                ],
                "summary": "Dado un id de publicación existente, elimina dicha publicación.",
                "description": "Dado un id de publicación existente, elimina dicha publicación.",
                "operationId": "App\\Http\\Controllers\\Api\\PublicationDataController::delete",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el mensaje: 'Publicació eliminada correctament'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'No existeix la publicació a eliminar'."
                    }
                }
            }
        },
...
    },
...
}